+++ docs/changelog.rst
@@ -15,6 +15,7 @@ Features
 Bugfixes
 ~~~~~~~~
 - `#9171 <https://0xacab.org/leap/bitmask-dev/issues/9171>`_: fix a bug in bootstrap that avoided more than one user to login.
+- `#9165 <https://0xacab.org/leap/bitmask-dev/issues/9165>`_: deprecate pyqt5-webkit, use qtwebengine instead.
 
 Misc
 ~~~~